+++ src/user/mod.rs
@@ -258,6 +258,34 @@ pub fn friends_no_retweets<'a>(con_token: &'a auth::Token, access_token: &'a aut
     parse_response(&mut resp)
 }
 
+///Lookup the relations between the authenticated user and the given accounts.
+pub fn relation_lookup(accts: &[UserID], con_token: &auth::Token, access_token: &auth::Token)
+    -> Result<Response<Vec<RelationLookup>>, error::Error>
+{
+    let mut params = HashMap::new();
+    let id_param = accts.iter()
+                        .filter_map(|x| match x {
+                            &UserID::ID(id) => Some(id.to_string()),
+                            _ => None,
+                        })
+                        .collect::<Vec<_>>()
+                        .join(",");
+    let name_param = accts.iter()
+                          .filter_map(|x| match x {
+                              &UserID::ScreenName(name) => Some(name),
+                              _ => None,
+                          })
+                          .collect::<Vec<_>>()
+                          .join(",");
+
+    add_param(&mut params, "user_id", id_param);
+    add_param(&mut params, "screen_name", name_param);
+
+    let mut resp = try!(auth::get(links::users::FRIENDSHIP_LOOKUP, con_token, access_token, Some(&params)));
+
+    parse_response(&mut resp)
+}
+
 ///Follow the given user with the authenticated account, and set whether device notifications
 ///should be enabled.
 ///