+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - New function to see friendship status between users (`relation`)
 - New function to change notification/retweet settings (`update_follow`)
 - New function to list users that the user has disabled retweets from (`friends_no_retweets`)
+- New function to look up friendship status for several users (`relation_lookup`)
 
 ### Changed
 - Combined IDLoader and UserLoader into CursorIter (with the same interface)